Handle null values in XML as well
This is the same as what we do for JSON, we make it possible to leave
out values that are null, which is important for certain special values
like UAString, where an empty value is different from a null value.

I considered making this a separate trait, but it would be annoying to
have to derive a trait that does nothing most of the time, and without
specialization there's no better way. We also can't put it on
BinaryEncodable since that trait isn't implemented for `Option`, and
can't really be at the moment. It's now effectively duplicated between
JsonEncodable and XmlEncodable, which is fine.