Support NodeSet2 xml files for codegen
Technically, BSD files are deprecated, and some standards don't define
them. What we're supposed to do instead is use NodeSet2 xml files. This
has some advantages, but also the major disadvantage that it is _really_
complicated.

This adds support then runs all our codegen in the new way. An advantage
is that we get docstrings, we could get more but I will save that for a
future PR.

An interesting side-effect of this is that the BSD schema refers only to
_actual_ types, i.e. how the types are represented in binary. The XML
schema refers to their real types, which in some cases are more high
level. For most of these I simply add a type alias (which may or may not
actually be sufficient for XML...), but for NumericRange specifically I
transition to using that type directly in the schema, which is quite
neat.

We also lose a few abstract types, which seems to fix a few bugs,
actually.