Use tracing instead of log in the library
This lays the foundation for much more thorough tracing support in the
main libraries. Tracing supports more or less drop-in support for log,
just replacing any use of the log macros with the same tracing macros.
Using the `log` feature, we can also emit tracing events as log events,
so consumers can consume them via normal logging. Hence, this should not
be a breaking change. We are still producing the same log events as
before.

Next up is actually using tracing to properly trace methods on the
server and client, both informative tracing and timing traces.