Refactor building server connector
This is necessary if we want to let users swap out the connector. The
final implementation is moderately hairy, because we connection
establishement has several stages:

 - Fetch endpoint URLs from the server _or_ use an explicit endpoint.
 - Create a reusable TCP connector.
 - Connect to the server.

Because we want old methods on the clinet to keep working, I add layer
in-between the new `ConnectionSource` trait, and the `Connector`, which
is essentially just `TryInto<Arc<dyn Connector>>`, `ConnectorBuilder`.

This means that the flow is

 - `ConnectionSource` accepts an endpoint URL and produces a
   `ConnectorBuilder`.
 - `ConnectorBuilder` creates a `Connector`.
 - `Connector` is reused each time we need to re-connect to create a
   `TcpConnection`.

In the future we might even want to be able to create a different type
of `Connection`, but I would want us to look into how we can ensure that
is still efficient, i.e. avoid boxing futures too much.