Run some services concurrently
This implements a general way to run some of our services concurrently,
notably _just_ the ones that are supposed to be run on one node manager
specifically, like `Read` and `Write`.

This involved some slightly fiddly async code, but the result isn't too
bad, and allows us to generalize a few things that we're not very
consistent about, like setting `current_node_manager_index` between
calls.

Doing this properly required cloning the `RequestContext`. I'm concerned
about the cost of this clone, which up until now we've been pretty lazy
about, so I moved the shared part into an `inner` field behind an `Arc`.
To avoid this breaking every server implementation I added a
`Deref<RequestContextInner>` to `RequestContext` which is a bad
practice, but does seem alright in this case.