Support for the OfType operator
IIRC I didn't do this originally because it was hard to fit into the
server architecture. We're a bit more settled now, and it is possible to
implement.

Still, it's a little annoying to do, since now evaluating event filters
requires the type tree, meaning that we need to share that somehow. I
made it so that SessionSubscriptions holds a reference to a type tree
for the user it belongs to, which _should_ be sufficient. It does mean
we take a read lock on the type tree whenever we notify about
subscriptions. The hope is that most servers will only rarely take a
write lock on the type tree, so this should have low contention.