Add fallback type loader
When running against unknown servers, it isn't really helpful that the
client fails if there is no registered type loader for a received type.
Instead, we'd like to continue, and ideally receive the value for
debugging or further processing.

This adds a fallback type loader which runs after all the other type
loaders and simply stores the incoming type as a raw binary, JSON, or
XML value depending on the input format. This can, in some
circumstances, be serialized back into the stream. There _are_
exceptions, since there for example is no standard way to encode JSON
extension object bodies in a binary stream.

This required a few changes to the main decoding interface:

 - We need the length of the extension object body when decoding binary
   bodies. This value isn't actually available in a few specific cases,
like in the custom type loader, meaning that you can't really combine
this with the custom type loader so that custom structures have _fields_
that lack a type loader. I think this is unlikely to be a big problem.
 - The XML decoder needs the name of the value, to be able to serialize
   it back later. This has a small performance cost, but I don't think
it's significant, and the XML decoder isn't that critical for
performance.

The fallback type loader is added to the decoding context by default,
with the lowest possible priority.