Forward fatal errors from subscription loop
If we get an error like `BadSessionClosed` or `BadSessionIdInvalid` from
the subscription event loop, it _usually_ means that we are already in
the process of closing the session.

In some cases it may, however, be the first indication of session
failure, so we can forward it to the main session event loop and trigger
session shutdown. This will reconnect according to normal rules.

The server would have to be _very_ misbehaved to respond to _anything_ with
`BadSessionClosed` or `BadSessionIdInvalid` and disconnecting being the
wrong response on our end.