Use is-empty or is-null-or-empty in a few places
Some clients poorly distinguish between empty strings and null strings,
this is a recurring problem. Prefer using `is_empty` for `UAString` and
`is_null_or_empty` for `ByteString`. It's very annoying that `is_empty`
has different semantics for `UAString` and `ByteString`, but it would be
a very breaking change to align them now.