Be more careful about adding timestamps
This can in some cases panic, see #148.

The case in that issue doesn't really make any sense to me. I don't
think the check we do there makes any sense.

That said, we do add `sampling_interval` in a few other places, we
should be careful about not overflowing and causing a panic.