Mechanisms to use NodeIdRef for notifications
This really is the main motivating example for the entire NodeIdRef
project. I was annoyed at how I was forced to create `NodeId`s for tags
in the `node-managers` example, while everywhere else I just refer to
them by their string value.

The result is reasonably easy to use, but gives great flexibility for
submitting "batches" of notifications. The old methods are still there,
just implemented via the new mechanism.