Client support for reverse connect
This implements the OPC-UA reverse connection protocol in the client
only, with a test using the .NET server.

While working with this I hit a panic in the secure channel, which we
should avoid, so I fixed it. Generally, this is not a very complex
implementation. To make it possible to provide a `:0` address, I let
users provide the TCP listener to the reverse connect source, which
works fine.