Server support for reverse connect
Implements support in the server SDK for the OPC-UA reverse connect
protocol, like we've already added to the server. To use it, you need to
register a reverse connect target for the server. The server will
periodically attempt to connect to the target and send a `ReverseHello`
message to establish a connection.

I ended up letting this be enabled by default, mostly because it
involves no extra dependencies, and the runtime cost is effectively
zero. It does nothing at all unless you register a reverse connect
target.