Clippy and lock file updates
New rust version, new clippy lints. Also updates the lockfile, and fixes
a few warnings caused by us being stuck at an old version of
generic-array. I expect we'll eventually be able to upgrade, but we have
that dependency through a few transitive dependencies, so it might take
a little while.

There is also a future-incompat warning from num-bigint-dig. We only
depend on this through the `rsa` crate, which _is_ maintained, and I
believe is replacing it with something rust-crypto maintained soon, so
that is unlikely to be a big issue.