Support dependent nodesets for types codegen
This is a relatively simple implementation of this. Let users list
dependent nodesets with import paths, like we allow for events. When
actually generating the types, we store this as a map from namespace to
import path, and during loading we just get the namespace of the type
being loaded.

This only really works for nodeset2 files, but since there exists a
workaround, this isn't critical. We can improve on this in the future.

With the codegen tests we can actually test stuff like this, which is
quite nice.