Avoid write lock on channel in client
This came with a bit of a refactor. In practice, we want to behave
slightly differently on the client and server, because the client is
always the party that _decides_ the security policy, so setting security
policy based on an open secure channel request only really makes sense
on the server.